Тесты оптимизированы
