+++ tests/conftest.py
@@ -1,0 +1,48 @@
+import datetime
+from collections import namedtuple
+
+import pytest
+import structlog
+
+from helpers.account_helper import AccountHelper
+from restclient.configation import Configuration as DmApiConfiguration
+from restclient.configation import Configuration as MailhogConfiguration
+from services.api_mailhog import MailHogApi
+from services.dm_api_account import DMApiAccount
+
+structlog.configure(
+    processors=[
+        structlog.processors.JSONRenderer(indent=4, ensure_ascii=True, sort_keys=True)
+    ]
+)
+
+
+@pytest.fixture
+def mailhog_api():
+    mailhog_configuration = MailhogConfiguration(host='http://5.63.153.31:5025')
+    mailhog_client = MailHogApi(configuration=mailhog_configuration)
+    return mailhog_client
+
+
+@pytest.fixture
+def account_api():
+    dm_api_configuration = DmApiConfiguration(host='http://5.63.153.31:5051', disable_log=False)
+    account = DMApiAccount(configuration=dm_api_configuration)
+    return account
+
+
+@pytest.fixture
+def account_helper(account_api, mailhog_api):
+    account_helper = AccountHelper(dm_account_api=account_api, mailhog=mailhog_api)
+    return account_helper
+
+@pytest.fixture
+def prepare_user():
+    now = datetime.datetime.now()
+    data = now.strftime("%d_%m_%Y_%H_%M_%S")
+    login = f'AlexBlok_{data}'
+    password = 'abcd12345'
+    email = f'{login}@mail.ru'
+    User = namedtuple("user", ["login", "password", "email"])
+    user = User(login=login, password=password, email=email)
+    return user
