+++ tests/function/put_v1_account_email/test_put_v1_account_email.py
@@ -1,30 +1,7 @@
-import structlog
-
-from helpers.account_helper import AccountHelper
-from restclient.configation import Configuration as DmApiConfiguration
-from restclient.configation import Configuration as MailhogConfiguration
-from services.api_mailhog import MailHogApi
-from services.dm_api_account import DMApiAccount
-
-structlog.configure(
-    processors=[
-        structlog.processors.JSONRenderer(indent=4, ensure_ascii=True, sort_keys=True)
-    ]
-)
-
-
-def test_post_v1_account_email():
-    mailhog_configuration = MailhogConfiguration(host='http://5.63.153.31:5025')
-    dm_api_configuration = DmApiConfiguration(host='http://5.63.153.31:5051', disable_log=False)
-
-    account = DMApiAccount(configuration=dm_api_configuration)
-    mailhog = MailHogApi(configuration=mailhog_configuration)
-
-    account_helper = AccountHelper(dm_account_api=account, mailhog=mailhog)
-
-    login = 'Rasputin7'
-    email = f'{login}@mail.ru'
-    password = '1234567'
+def test_post_v1_account_email(account_helper,prepare_user):
+    login = prepare_user.login
+    password = prepare_user.password
+    email = prepare_user.email
 
     account_helper.register_new_user(login=login, password=password, email=email)
     account_helper.user_login(login=login, password=password)
